Switch to yt-dlp, use RSS for IDS, drop jq
This commit drops support for youtube-dl so that channel IDs can be
found with yt-dlp specific features. This is then used for obtaining a
channels RSS feed (eg.
https://www.youtube.com/feeds/videos.xml?channel_id=UC2eYFnH61tmytImy1mTYvhA),
which contains recent channel video IDs.

This is better than scraping the /videos page of a channel due to the
unstable nature of scraping since a page layout rework will require a
rework of this project. RSS, however, is much easier to parse and much
less likely to have its layout modified in the future.